Optimize rubik image build (#78)
This removes various apt calls, poorly implemented/unnecessary function,
and contains various optimizations to improve build speed for the
rubikpi3 image, and the rubikpi3 image mounting process. Using the
`mount_rubikpi3.sh` script, our runtime on
https://github.com/photonvision/photonvision is functionally equivalent
to that of an opi image, and we've reached ~55m runs in this repo.

About 10 minutes of that time is spent reaching the thundercomm apt ppa,
if we want to speed it up we could consider mirroring or proxy-caching.
We also spend some time in qemu, exploring an arm-runner could
potentially alleviate that.