Native-arm-runner-action (#97)
Utilize the
[photon-image-runner](https://github.com/PhotonVision/photon-image-runner)
action to build images. Once this is validated, we can use the same
action for modifying the images in
[photonvision](https://github.com/PhotonVision/photonvision) and
[photon-libcamera-gl-driver](https://github.com/PhotonVision/photon-libcamera-gl-driver)
workflows.

Improvements in this PR:
* Like #95, this uses the GitHub arm runner to greatly speed up image
build times.
* Images specified by a yaml file, like the RubikPi3, are handled
correctly.
* A PhotonVision release can be specified in the workflow, which allows
images to be built with releases other than latest.
* The action implements an option to specify the minimum free space
needed in the image. This should prevent repeated increases in the size
of an image if it can't be shrunk by the action.
* Several minor problems were fixed, including
  * CIDATA partition is properly mounted on OrangePi
  * Version information is added by `install_common.sh`
  * Empty space is zero filled on all images for better compression
  * No more log spam from Dpkg

Supersedes #95